+++ src/main/java/org/prgrms/kdt/domain/voucher/model/PercentDiscountVoucher.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 public class PercentDiscountVoucher implements Voucher {
 
     private final UUID voucherId;
-    private final long percent;
+    private final int percent;
 
-    public PercentDiscountVoucher(UUID voucherId, long percent) {
+    public PercentDiscountVoucher(UUID voucherId, int percent) {
         this.voucherId = voucherId;
         this.percent = percent;
     }
